Add projects.
