+++ crates/cli/src/commands/sync.rs
@@ -1,32 +1,14 @@
-use crate::helpers::create_progress_bar;
-use moon::{build_dep_graph, generate_project_graph, load_workspace};
-use moon_action_pipeline::Pipeline;
+use moon_logger::warn;
 use starbase::AppResult;
+use starbase_styles::color;
 
 pub async fn sync() -> AppResult {
-    let done = create_progress_bar("Syncing projects...");
-
-    let mut workspace = load_workspace().await?;
-    let project_graph = generate_project_graph(&mut workspace).await?;
-    let mut project_count = 0;
-    let mut dep_builder = build_dep_graph(&workspace, &project_graph);
-
-    for project in project_graph.get_all()? {
-        dep_builder.sync_project(project)?;
-        project_count += 1;
-    }
-
-    let dep_graph = dep_builder.build();
-
-    let mut pipeline = Pipeline::new(workspace, project_graph);
-    let results = pipeline.run(dep_graph, None).await?;
-
-    done(
-        format!("Successfully synced {project_count} projects"),
-        true,
+    warn!(
+        "This command is deprecated. Use {} instead.",
+        color::shell("moon sync projects")
     );
 
-    pipeline.render_results(&results)?;
+    crate::commands::syncs::projects::sync().await?;
 
     Ok(())
 }
