+++ crates/cli/src/commands/syncs/mod.rs
@@ -1,1 +1,2 @@
 pub mod codeowners;
+pub mod projects;
